Move operation default helpers to helpers/operationDefaults.ts
Relocate folderIdForOperation, toolMatchesTemplate, operationKindLabel,
isOperationTargetValid, defaultOperationName, defaultOperationForTarget,
and fallbackOperationTarget into a dedicated home module. Slices now
import these directly.
